+++ changelog.md
@@ -1,12 +1,19 @@
 <h1> Changelog </h1>
 
-> Qrala _sine 2021_ <br/> 
-> Version: 0.9.2 <br/> 
+> Qrala &bull; _since 2021_ &bull;  <br/> 
+> Version: 0.9.4 <br/> 
 > All rights reserved. <br/>
 > Copyright &copy; 2021 - 2022. [Christoph Rohde](https://github.com/CodebyCR) <br/>
 > Licence: GNU Affero General Public License v3.0
 #
 
+---
+<h3>Version 0.9.4)  </h3>
+
+- add .gitignore
+- add some code for VCard QR-Codes
+- add some code for WIFI QR-Codes
+- some readability improvements
 
 
 ---